Move implementation over
