+++ pyvrp/cpp/LoadSegment.h
@@ -75,11 +75,11 @@ public:
     LoadSegment() = default;  // default is all zero
 
     // Construct from load attributes of the given client and dimension.
-    LoadSegment(ProblemData::Client const &client, size_t dimension);
+    LoadSegment(Client const &client, size_t dimension);
 
     // Construct from initial load attributes of the given vehicle type and
     // dimension.
-    LoadSegment(ProblemData::VehicleType const &vehicleType, size_t dimension);
+    LoadSegment(VehicleType const &vehicleType, size_t dimension);
 
     // Construct from raw data.
     inline LoadSegment(Load delivery,
